default.xml: add west repository in build/other/west
West has grown up into its own repository, which we need to continue
to use as the real implementation for zmp flash and zmp debug. Clone
it into a hidden directory for now as we're still pretending those zmp
features are standalone. This directory will move into west/west (next
to west/manifest) when we start using west's multirepo features.

Change-Id: I69ba2ec6c7ec64ef38c555755321b3d0c072ea95
Signed-off-by: Marti Bolivar <marti@foundries.io>