Go west
Add a west manifest, west.yml, and a requirements.txt for the
ZmP (been wanting to do that one for a while).

The west manifest contains neither the dm-lwm2m-light repository nor
the prebuilt toolchains. The former was too much of an internal app to
continue to distribute. The latter lack a cross-platform solution;
users will have to install their own cross toolchain using help in the
documentation.

Drop the old repo manifest, leaving behind a readme explaining what
happened.

Signed-off-by: Marti Bolivar <marti@foundries.io>